[SDK-109] Changed some CloudConstant names and changed the structure of the sample-client
