+++ njams-sdk-communication-cloud/src/main/java/com/im/njams/sdk/communication/cloud/CloudConstants.java
@@ -33,4 +33,33 @@ public class CloudConstants {
     public static final String CLIENT_PRIVATEKEY = PROPERTY_PREFIX + ".client.privatekey";
     
     public static final String MAX_PAYLOAD_BYTES = PROPERTY_PREFIX + ".maxPayloadBytes";
+
+    public static final String CLIENT_KS = "client.ks";
+
+    public static final String CLIENT_TS = "client.ts";
+
+    public static final String TRUST_STORE = "javax.net.ssl.trustStore";
+
+    /**
+     * Content type json
+     */
+    public static final String CONTENT_TYPE_JSON = "application/json";
+
+    /**
+     * Content type text plain
+     */
+    public  static final String CONTENT_TYPE = "Content-Type";
+    public  static final String CONTENT_TYPE_TEXT = "text/plain";
+
+    public  static final String HTTP_REQUEST_TYPE_POST = "POST";
+    public  static final String HTTP_REQUEST_TYPE_PUT= "PUT";
+    public  static final String HTTP_REQUEST_TYPE_GET= "GET";
+
+    public  static final String ACCEPT = "Accept";
+
+    public  static final String CONTENT_LANGUAGE = "Content-Language";
+
+    public  static final String CONTENT_LANGUAGE_EN_US = "en-US";
+
+    public  static final String UTF_8 = "charset=UTF-8";
 }
